Extract runtime startup orchestration into agent/runtime/startup.rs
